+++ src/utils/cpp_utils/rt_funcs/vindex.cpp
@@ -2,14 +2,15 @@
         value_t iterable = iterable_v.get_value();
         value_t index = index_v.get_value();
 
-        if (!std::holds_alternative<int>(index)){
+        if (!std::holds_alternative<long long>(index))
+        {
             RunTime instance;
             instance.throw_rt_error("Indexing expressions must be of type 'int'");
             return Value(none{});
         }
 
         if (std::holds_alternative<std::vector<Value> >(iterable)){
-            int int_index = std::get<int>(index);
+            long long int_index = std::get<long long>(index);
 
             if (int_index >= std::get<std::vector<Value> >(iterable).size() | int_index < 0){
                 RunTime instance;
@@ -19,7 +20,7 @@
             return std::get<std::vector<Value> >(iterable)[int_index];
         }
         else if (std::holds_alternative<std::string>(iterable)){
-            int int_index = std::get<int>(index);
+            long long int_index = std::get<long long>(index);
 
             if (int_index >= std::get<std::string>(iterable).length() | int_index < 0){
                 RunTime instance;
